semester 5/DBs/Laboratory work 1/pyproject.toml
[project]
name = "laboratory-work-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0.44",
    "fastapi>=0.115.2",
    "uvicorn>=0.30.6",
    "asyncpg>=0.29.0",
    "jinja2>=3.1.4",
    "python-dotenv>=1.0.1",
    "psycopg>=3.2.12",
    "python-multipart>=0.0.20",
    "greenlet>=3.2.4",
]
